api ëë리팩토링
