--- backend/src/main/java/com/grepp/smartwatcha/app/controller/api/recommend/RecommendLatestApiController.java
@@ -1,23 +1,0 @@
-package com.grepp.smartwatcha.app.controller.api.recommend;
-
-import com.grepp.smartwatcha.app.model.recommend.RecommendLatestMovieService;
-import com.grepp.smartwatcha.app.controller.api.recommend.payload.MovieRecommendLatestResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/recommend")
-public class RecommendLatestApiController {
-
-    private final RecommendLatestMovieService recommendService;
-
-    @GetMapping("/latest-movies")
-    public List<MovieRecommendLatestResponse> getLatestMovies() {
-        return recommendService.getTop10LatestMovies();
-    }
-}
